+++ ios/sdk/component/listview/HippyBaseListView.m
@@ -340,13 +340,6 @@
 
 #pragma mark - Scroll
 
-#define MTT_FORWARD_SCROLL_EVENT(call) \
-for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) { \
-if ([scrollViewListener respondsToSelector:_cmd]) { \
-[scrollViewListener call]; \
-} \
-}
-
 - (void)scrollViewDidScroll:(UIScrollView *)scrollView
 {
 	NSTimeInterval now = CACurrentMediaTime();
@@ -357,7 +350,11 @@ if ([scrollViewListener respondsToSelector:_cmd]) { \
 		_lastScrollDispatchTime = now;
 	}
 	
-	MTT_FORWARD_SCROLL_EVENT(scrollViewDidScroll: scrollView);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewDidScroll:)]) {
+            [scrollViewListener scrollViewDidScroll:scrollView];
+        }
+    }
 }
 
 - (void)scrollViewWillBeginDragging:(UIScrollView *)scrollView
@@ -365,9 +362,13 @@ if ([scrollViewListener respondsToSelector:_cmd]) { \
 	if (self.onScrollBeginDrag) {
 		self.onScrollBeginDrag([self scrollBodyData]);
 	}
-  _manualScroll = YES;
+    _manualScroll = YES;
 	[self cancelTouch];
-	MTT_FORWARD_SCROLL_EVENT(scrollViewWillBeginDragging: scrollView);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewWillBeginDragging:)]) {
+            [scrollViewListener scrollViewWillBeginDragging:scrollView];
+        }
+    }
 }
 
 - (void)scrollViewWillBeginDecelerating:(UIScrollView *)scrollView
@@ -375,7 +376,11 @@ if ([scrollViewListener respondsToSelector:_cmd]) { \
 	if (self.onMomentumScrollBegin) {
 		self.onMomentumScrollBegin([self scrollBodyData]);
 	}
-	MTT_FORWARD_SCROLL_EVENT(scrollViewWillBeginDecelerating: scrollView);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewWillBeginDecelerating:)]) {
+            [scrollViewListener scrollViewWillBeginDecelerating:scrollView];
+        }
+    }
 }
 
 - (void)scrollViewDidEndDragging:(UIScrollView *)scrollView willDecelerate:(BOOL)decelerate
@@ -383,7 +388,11 @@ if ([scrollViewListener respondsToSelector:_cmd]) { \
 	if (!decelerate) {
 		_manualScroll = NO;
 	}
-	MTT_FORWARD_SCROLL_EVENT(scrollViewDidEndDragging: scrollView willDecelerate: decelerate);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewDidEndDragging:willDecelerate:)]) {
+            [scrollViewListener scrollViewDidEndDragging:scrollView willDecelerate:decelerate];
+        }
+    }
 }
 
 - (void)scrollViewWillEndDragging:(UIScrollView *)scrollView withVelocity:(CGPoint)velocity targetContentOffset:(inout CGPoint *)targetContentOffset
@@ -399,7 +408,11 @@ if ([scrollViewListener respondsToSelector:_cmd]) { \
 		self.onScrollEndDrag([self scrollBodyData]);
 	}
 	
-	MTT_FORWARD_SCROLL_EVENT(scrollViewWillEndDragging: scrollView withVelocity: velocity targetContentOffset: targetContentOffset);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewWillEndDragging:withVelocity:targetContentOffset:)]) {
+            [scrollViewListener scrollViewWillEndDragging:scrollView withVelocity:velocity targetContentOffset:targetContentOffset];
+        }
+    }
 }
 
 - (void)scrollViewDidEndDecelerating:(UIScrollView *)scrollView
@@ -412,27 +425,47 @@ if ([scrollViewListener respondsToSelector:_cmd]) { \
 		self.onMomentumScrollEnd([self scrollBodyData]);
 	}
 	
-	MTT_FORWARD_SCROLL_EVENT(scrollViewDidEndDecelerating:scrollView);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewDidEndDecelerating:)]) {
+            [scrollViewListener scrollViewDidEndDecelerating:scrollView];
+        }
+    }
 }
 
 - (void)scrollViewWillBeginZooming:(UIScrollView *)scrollView withView:(UIView *)view
 {
-	MTT_FORWARD_SCROLL_EVENT(scrollViewWillBeginZooming: scrollView withView: view);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewWillBeginZooming:withView:)]) {
+            [scrollViewListener scrollViewWillBeginZooming:scrollView withView:view];
+        }
+    }
 }
 
 - (void)scrollViewDidZoom:(UIScrollView *)scrollView
 {
-	MTT_FORWARD_SCROLL_EVENT(scrollViewDidZoom:scrollView);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewDidZoom:)]) {
+            [scrollViewListener scrollViewDidZoom:scrollView];
+        }
+    }
 }
 
 - (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale
 {
-	MTT_FORWARD_SCROLL_EVENT(scrollViewDidEndZooming: scrollView withView: view atScale: scale);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewDidEndZooming:withView:atScale:)]) {
+            [scrollViewListener scrollViewDidEndZooming:scrollView withView:view atScale:scale];
+        }
+    }
 }
 
 - (void)scrollViewDidEndScrollingAnimation:(UIScrollView *)scrollView
 {
-	MTT_FORWARD_SCROLL_EVENT(scrollViewDidEndScrollingAnimation: scrollView);
+    for (NSObject<UIScrollViewDelegate> *scrollViewListener in _scrollListeners) {
+        if ([scrollViewListener respondsToSelector:@selector(scrollViewDidEndScrollingAnimation:)]) {
+            [scrollViewListener scrollViewDidEndScrollingAnimation:scrollView];
+        }
+    }
 }
 
 - (NSDictionary *)scrollBodyData