+++ ios/sdk/utils/NSDictionary+HippyDictionaryDeepCopy.h
@@ -23,10 +23,6 @@
 #import <Foundation/Foundation.h>
 #import "HippyDeepCopyProtocol.h"
 
-NS_ASSUME_NONNULL_BEGIN
-
 @interface NSDictionary (HippyDictionaryDeepCopy)<HippyDeepCopyProtocol>
 
 @end
-
-NS_ASSUME_NONNULL_END