+++ scripts/react-configs.js
@@ -70,6 +70,7 @@ const builds = {
         'bezier-easing',
         'debounce',
         'swiper',
+        'rmc-list-view',
       ].find(ext => id.startsWith(ext));
     },
   },
@@ -83,6 +84,7 @@ function genConfig(name) {
     plugins: [
       replace({
         'process.env.HIPPY_REACT_VERSION': `"${hippyReactPackage.version}"`,
+        'process.env.HIPPY_REACT_WEB_VERSION': `"${hippyReactWebPackage.version}"`,
       }),
       typescript({
         transpileOnly: true,