+++ types/event.ts
@@ -1,31 +1,33 @@
 /* eslint-disable import/prefer-default-export */
 /* eslint-disable camelcase */
 
+interface LayoutContent {
+  /**
+   * The position X of component
+   */
+  x: number;
+
+  /**
+   * The position Y of component
+   */
+  y: number;
+
+  /**
+   * The width of component
+   */
+  width: number;
+
+  /**
+   * The height of component
+   */
+  height: number;
+}
+
 interface LayoutEvent {
   /**
    * The event data of layout event
    */
-  nativeEvent: {
-    /**
-     * The position X of component
-     */
-    x: number;
-
-    /**
-     * The position Y of component
-     */
-    y: number;
-
-    /**
-     * The width of component
-     */
-    width: number;
-
-    /**
-     * The height of component
-     */
-    height: number;
-  };
+  nativeEvent: LayoutContent;
 }
 
 interface TouchEvent {
@@ -49,6 +51,7 @@ interface FocusEvent {
 }
 
 export {
+  LayoutContent,
   LayoutEvent,
   TouchEvent,
   FocusEvent,